Llms: remove the 'types' file and extract the vendor description out
