+++ src/apps/models-modal/ModelsSourceSelector.tsx
@@ -8,23 +8,23 @@ import CloudOutlinedIcon from '@mui/icons-material/CloudOutlined';
 import ComputerIcon from '@mui/icons-material/Computer';
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 
-import { DModelSourceId, ModelVendor, ModelVendorId } from '~/modules/llms/llm.types';
-import { createModelSourceForVendor, findAllVendors, findVendorById } from '~/modules/llms/vendor.registry';
+import { DModelSourceId, useModelsStore } from '~/modules/llms/store-llms';
+import { IModelVendor, ModelVendorId } from '~/modules/llms/vendors/IModelVendor';
+import { createModelSourceForVendor, findAllVendors, findVendorById } from '~/modules/llms/vendors/vendor.registry';
 import { hasServerKeyOpenAI } from '~/modules/llms/openai/openai.vendor';
-import { useModelsStore } from '~/modules/llms/store-llms';
 
 import { CloseableMenu } from '~/common/components/CloseableMenu';
 import { ConfirmationModal } from '~/common/components/ConfirmationModal';
 import { hideOnDesktop, hideOnMobile } from '~/common/theme';
 
 
-function locationIcon(vendor?: ModelVendor | null) {
+function locationIcon(vendor?: IModelVendor | null) {
   if (vendor && vendor.id === 'openai' && hasServerKeyOpenAI)
     return <CloudDoneOutlinedIcon />;
   return !vendor ? null : vendor.location === 'local' ? <ComputerIcon /> : <CloudOutlinedIcon />;
 }
 
-function vendorIcon(vendor?: ModelVendor | null) {
+function vendorIcon(vendor?: IModelVendor | null) {
   const Icon = !vendor ? null : vendor.Icon;
   return Icon ? <Icon /> : null;
 }
