+++ src/apps/models-modal/VendorSourceSetup.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
-import { DModelSource } from '~/modules/llms/llm.types';
-import { findVendorById } from '~/modules/llms/vendor.registry';
+import { DModelSource } from '~/modules/llms/store-llms';
+import { findVendorById } from '~/modules/llms/vendors/vendor.registry';
 
 
 export function VendorSourceSetup(props: { source: DModelSource }) {
