+++ src/apps/personas/useLLMChain.ts
@@ -1,8 +1,7 @@
 import * as React from 'react';
 
-import { DLLMId } from '~/modules/llms/llm.types';
+import { DLLMId, useModelsStore } from '~/modules/llms/store-llms';
 import { callChatGenerate, VChatMessageIn } from '~/modules/llms/llm.client';
-import { useModelsStore } from '~/modules/llms/store-llms';
 
 
 export interface LLMChainStep {
