+++ src/common/util/token-counter.ts
@@ -1,8 +1,7 @@
 import { encoding_for_model, get_encoding, Tiktoken, TiktokenModel } from '@dqbd/tiktoken';
 
-import { DLLMId } from '~/modules/llms/llm.types';
 import { findLLMOrThrow } from '~/modules/llms/llm.client';
-import { useModelsStore } from '~/modules/llms/store-llms';
+import { DLLMId, useModelsStore } from '~/modules/llms/store-llms';
 
 
 // Do not set this to true in production, it's very verbose
