+++ src/modules/aifn/react/react.ts
@@ -2,7 +2,7 @@
  * porting of implementation from here: https://til.simonwillison.net/llms/python-react-pattern
  */
 
-import { DLLMId } from '~/modules/llms/llm.types';
+import { DLLMId } from '~/modules/llms/store-llms';
 import { callApiSearchGoogle } from '~/modules/google/search.client';
 import { callChatGenerate, VChatMessageIn } from '~/modules/llms/llm.client';
