+++ src/modules/llms/openrouter/OpenRouterSourceSetup.tsx
@@ -12,8 +12,7 @@ import { InlineError } from '~/common/components/InlineError';
 import { Link } from '~/common/components/Link';
 import { settingsGap } from '~/common/theme';
 
-import { DLLM, DModelSource, DModelSourceId } from '../llm.types';
-import { useModelsStore, useSourceSetup } from '../store-llms';
+import { DLLM, DModelSource, DModelSourceId, useModelsStore, useSourceSetup } from '../store-llms';
 
 import { isValidOpenRouterKey, ModelVendorOpenRouter } from './openrouter.vendor';
